+++ src/core/api/public_api.ts
@@ -96,11 +96,7 @@ import initializeMediaSourcePlayback, {
   IReloadingMediaSourceEvent,
   IStalledEvent,
 } from "../init";
-import {
-  IPublicNonFiniteStreamEvent,
-  IPublicStreamEvent,
-  IStreamEvent,
-} from "../init/stream_events_emitter";
+import { IStreamEventData } from "../init/stream_events_emitter";
 import SourceBuffersStore, {
   IBufferedChunk,
   IBufferType,
@@ -171,6 +167,13 @@ interface IBitrateEstimate {
   bitrate : number | undefined;
 }
 
+export type IStreamEvent = { data: IStreamEventData;
+                             start: number;
+                             end: number;
+                             onExit?: () => void; } |
+                           { data: IStreamEventData;
+                             start: number; };
+
 /** Every events sent by the RxPlayer's public API. */
 interface IPublicAPIEvent {
   playerStateChange : string;
@@ -199,8 +202,8 @@ interface IPublicAPIEvent {
                                   representation : Representation; }>;
   seeking : null;
   seeked : null;
-  streamEvent : IPublicStreamEvent|IPublicNonFiniteStreamEvent;
-  streamEventSkip : IPublicStreamEvent|IPublicNonFiniteStreamEvent;
+  streamEvent : IStreamEvent;
+  streamEventSkip : IStreamEvent;
 }
 
 /**
@@ -2023,10 +2026,10 @@ class Player extends EventEmitter<IPublicAPIEvent> {
   private _priv_onPlaybackEvent(event : IInitEvent) : void {
     switch (event.type) {
       case "stream-event":
-        this._priv_onStreamEvent(event);
+        this.trigger("streamEvent", event.value);
         break;
       case "stream-event-skip":
-        this._priv_onStreamEventSkip(event);
+        this.trigger("streamEventSkip", event.value);
         break;
       case "activePeriodChanged":
         this._priv_onActivePeriodChanged(event.value);
@@ -2177,14 +2180,6 @@ class Player extends EventEmitter<IPublicAPIEvent> {
           this._priv_trackChoiceManager.update();
         }
       });
-  }
-
-  private _priv_onStreamEvent(event: IStreamEvent) {
-    this.trigger("streamEvent", event.value);
-  }
-
-  private _priv_onStreamEventSkip(event: IStreamEvent) {
-    this.trigger("streamEventSkip", event.value);
   }
 
   /**
@@ -2622,3 +2617,4 @@ class Player extends EventEmitter<IPublicAPIEvent> {
 Player.version = /*PLAYER_VERSION*/"3.21.0";
 
 export default Player;
+export { IStreamEventData };